refactor(app): extract core and channel composition roots with arch guards
cmd/agent/app.go splits into internal/app/core (shared domain and agent
runtime providers, including the new provideTurnService binding) and
internal/app/channel (the Channel boundary root: registry, manager,
inbound processor, discuss pipeline, email, webhook tunnel). cmd/agent
keeps only config, HTTP handlers, and the server. internal/arch tests
enforce the boundary import rules and confine team.DefaultTeamID to the
DB layer, composition roots, and documented resolver defaults.
