+++ cmd/agent/http_providers.go
@@ -1,0 +1,196 @@
+package main
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"log/slog"
+	"net/http"
+	"strings"
+
+	"go.uber.org/fx"
+
+	"github.com/memohai/memoh/internal/accounts"
+	"github.com/memohai/memoh/internal/acpagent"
+	"github.com/memohai/memoh/internal/agent/background"
+	appcore "github.com/memohai/memoh/internal/app/core"
+	audiopkg "github.com/memohai/memoh/internal/audio"
+	"github.com/memohai/memoh/internal/boot"
+	"github.com/memohai/memoh/internal/bots"
+	"github.com/memohai/memoh/internal/channel"
+	"github.com/memohai/memoh/internal/channel/adapters/local"
+	"github.com/memohai/memoh/internal/channel/route"
+	"github.com/memohai/memoh/internal/command"
+	"github.com/memohai/memoh/internal/config"
+	"github.com/memohai/memoh/internal/conversation"
+	"github.com/memohai/memoh/internal/conversation/flow"
+	dbstore "github.com/memohai/memoh/internal/db/store"
+	emailpkg "github.com/memohai/memoh/internal/email"
+	"github.com/memohai/memoh/internal/handlers"
+	"github.com/memohai/memoh/internal/healthcheck"
+	channelchecker "github.com/memohai/memoh/internal/healthcheck/checkers/channel"
+	mcpchecker "github.com/memohai/memoh/internal/healthcheck/checkers/mcp"
+	modelchecker "github.com/memohai/memoh/internal/healthcheck/checkers/model"
+	"github.com/memohai/memoh/internal/mcp"
+	"github.com/memohai/memoh/internal/media"
+	memprovider "github.com/memohai/memoh/internal/memory/adapters"
+	"github.com/memohai/memoh/internal/message"
+	"github.com/memohai/memoh/internal/message/event"
+	"github.com/memohai/memoh/internal/models"
+	"github.com/memohai/memoh/internal/oauthclients"
+	"github.com/memohai/memoh/internal/providers"
+	"github.com/memohai/memoh/internal/server"
+	sessionpkg "github.com/memohai/memoh/internal/session"
+	"github.com/memohai/memoh/internal/settings"
+	"github.com/memohai/memoh/internal/toolapproval"
+	"github.com/memohai/memoh/internal/userinput"
+	"github.com/memohai/memoh/internal/version"
+	"github.com/memohai/memoh/internal/workspace"
+)
+
+func provideServerHandler(fn any) any {
+	return fx.Annotate(
+		fn,
+		fx.As(new(server.Handler)),
+		fx.ResultTags(`group:"server_handlers"`),
+	)
+}
+
+func provideMemoryHandler(log *slog.Logger, botService *bots.Service, accountService *accounts.Service, _ config.Config, memoryRegistry *memprovider.Registry, settingsService *settings.Service, _ *handlers.ContainerdHandler) *handlers.MemoryHandler {
+	h := handlers.NewMemoryHandler(log, botService, accountService)
+	h.SetMemoryRegistry(memoryRegistry)
+	h.SetSettingsService(settingsService)
+	return h
+}
+
+func provideAuthHandler(log *slog.Logger, accountService *accounts.Service, rc *boot.RuntimeConfig) *handlers.AuthHandler {
+	return handlers.NewAuthHandler(log, accountService, rc.JwtSecret, rc.JwtExpiresIn)
+}
+
+func provideMessageHandler(log *slog.Logger, chatService *conversation.Service, msgService *message.DBService, sessionService *sessionpkg.Service, mediaService *media.Service, botService *bots.Service, accountService *accounts.Service, hub *event.Hub, toolApproval *toolapproval.Service, userInput *userinput.Service, bgManager *background.Manager) *handlers.MessageHandler {
+	h := handlers.NewMessageHandler(log, chatService, msgService, sessionService, botService, accountService, hub)
+	h.SetMediaService(mediaService)
+	h.SetToolApprovalService(toolApproval)
+	h.SetUserInputService(userInput)
+	h.SetBackgroundManager(bgManager)
+	return h
+}
+
+func provideSessionHandler(log *slog.Logger, sessionService *sessionpkg.Service, acpPool *acpagent.SessionPool, botService *bots.Service, accountService *accounts.Service) *handlers.SessionHandler {
+	return handlers.NewSessionHandler(log, sessionService, acpPool, botService, accountService)
+}
+
+func provideUsersHandler(log *slog.Logger, accountService *accounts.Service, botService *bots.Service, routeService *route.DBService, channelStore *channel.Store, channelLifecycle *channel.Lifecycle, channelManager *channel.Manager, registry *channel.Registry, workspaceManager *workspace.Manager, acpPool *acpagent.SessionPool) *handlers.UsersHandler {
+	handler := handlers.NewUsersHandler(log, accountService, botService, routeService, channelStore, channelLifecycle, channelManager, registry, workspaceManager)
+	handler.SetACPRuntimeCloser(acpPool)
+	return handler
+}
+
+func provideACPCodexOAuthServerHandler(handler *handlers.ACPCodexOAuthHandler) *handlers.ACPCodexOAuthHandler {
+	return handler
+}
+
+func provideACPClaudeCodeOAuthServerHandler(handler *handlers.ACPClaudeCodeOAuthHandler) *handlers.ACPClaudeCodeOAuthHandler {
+	return handler
+}
+
+func provideProviderOAuthHandler(providersService *providers.Service, acpCodexOAuthHandler *handlers.ACPCodexOAuthHandler) *handlers.ProviderOAuthHandler {
+	handler := handlers.NewProviderOAuthHandler(providersService)
+	handler.SetACPCodexOAuthHandler(acpCodexOAuthHandler)
+	return handler
+}
+
+func provideWebHandler(channelManager *channel.Manager, channelStore *channel.Store, chatService *conversation.Service, hub *local.RouteHub, botService *bots.Service, accountService *accounts.Service, sessionService *sessionpkg.Service, resolver *flow.Resolver, mediaService *media.Service, audioService *audiopkg.Service, settingsService *settings.Service, rc *boot.RuntimeConfig, commandHandler *command.Handler, containerdHandler *handlers.ContainerdHandler) *handlers.LocalChannelHandler {
+	h := handlers.NewLocalChannelHandler(local.WebType, channelManager, channelStore, chatService, hub, botService, accountService, sessionService)
+	h.SetResolver(resolver)
+	h.SetCommandHandler(commandHandler)
+	h.SetRuntimeSkillResolver(containerdHandler)
+	h.SetAuthTokenConfig(rc.JwtSecret, rc.JwtExpiresIn)
+	h.SetMediaService(mediaService)
+	h.SetSpeechService(audioService, &webSpeechModelResolver{settings: settingsService})
+	return h
+}
+
+func provideEmailOAuthHandler(log *slog.Logger, service *emailpkg.Service, tokenStore *emailpkg.DBOAuthTokenStore, oauthClients *oauthclients.Registry, cfg config.Config) *handlers.EmailOAuthHandler {
+	addr := strings.TrimSpace(cfg.Server.Addr)
+	if addr == "" {
+		addr = ":8080"
+	}
+	host := addr
+	if strings.HasPrefix(host, ":") {
+		host = "localhost" + host
+	}
+	callbackURL := "http://" + host + "/api/email/oauth/callback"
+	return handlers.NewEmailOAuthHandler(log, service, tokenStore, oauthClients, callbackURL)
+}
+
+type serverParams struct {
+	fx.In
+
+	Logger            *slog.Logger
+	RuntimeConfig     *boot.RuntimeConfig
+	Config            config.Config
+	ServerHandlers    []server.Handler `group:"server_handlers"`
+	ContainerdHandler *handlers.ContainerdHandler
+}
+
+func provideServer(params serverParams) *server.Server {
+	allHandlers := make([]server.Handler, 0, len(params.ServerHandlers)+1)
+	allHandlers = append(allHandlers, params.ServerHandlers...)
+	allHandlers = append(allHandlers, params.ContainerdHandler)
+	return server.NewServer(params.Logger, params.RuntimeConfig.ServerAddr, params.Config.Auth.JWTSecret, allHandlers...)
+}
+
+func startServer(lc fx.Lifecycle, logger *slog.Logger, srv *server.Server, shutdowner fx.Shutdowner, cfg config.Config, queries dbstore.Queries, accountStore dbstore.AccountStore, emailService *emailpkg.Service, botService *bots.Service, _ *handlers.ContainerdHandler, manager *workspace.Manager, mcpConnService *mcp.ConnectionService, toolGateway *mcp.ToolGatewayService, channelManager *channel.Manager, modelsService *models.Service) {
+	fmt.Printf("Starting Memoh Agent %s\n", version.GetInfo())
+
+	lc.Append(fx.Hook{
+		OnStart: func(ctx context.Context) error {
+			if err := appcore.EnsureAdminUser(ctx, logger, accountStore, emailService, cfg); err != nil {
+				return err
+			}
+			botService.SetContainerLifecycle(manager)
+			botService.SetContainerReachability(func(ctx context.Context, botID string) error {
+				_, err := manager.MCPClient(ctx, botID)
+				return err
+			})
+			botService.AddRuntimeChecker(healthcheck.NewRuntimeCheckerAdapter(
+				mcpchecker.NewChecker(logger, mcpConnService, toolGateway),
+			))
+			botService.AddRuntimeChecker(healthcheck.NewRuntimeCheckerAdapter(
+				channelchecker.NewChecker(logger, channelManager),
+			))
+			botService.AddRuntimeChecker(healthcheck.NewRuntimeCheckerAdapter(
+				modelchecker.NewChecker(logger, modelchecker.NewQueriesLookup(queries), modelsService),
+			))
+
+			go func() {
+				if err := srv.Start(); err != nil && !errors.Is(err, http.ErrServerClosed) {
+					logger.Error("server failed", slog.Any("error", err))
+					_ = shutdowner.Shutdown()
+				}
+			}()
+			return nil
+		},
+		OnStop: func(ctx context.Context) error {
+			if err := srv.Stop(ctx); err != nil && !errors.Is(err, http.ErrServerClosed) {
+				return fmt.Errorf("server stop: %w", err)
+			}
+			return nil
+		},
+	})
+}
+
+// webSpeechModelResolver adapts bot settings to the web chat speech model
+// lookup (same shape as appchannel's inbound resolver glue).
+type webSpeechModelResolver struct {
+	settings *settings.Service
+}
+
+func (r *webSpeechModelResolver) ResolveSpeechModelID(ctx context.Context, botID string) (string, error) {
+	s, err := r.settings.GetBot(ctx, botID)
+	if err != nil {
+		return "", err
+	}
+	return s.TtsModelID, nil
+}
